Fixed a cast issue with frame rates

+++ core/src/Controller/Controller.h
@@ -46,12 +46,12 @@ private:
     /**
      * How quickly the game logic updates. 1 over the number of updates per second.
      */
-    const double UPDATE_RATE = (1 / 60.0) * 1000;
+    const unsigned int UPDATE_RATE = (int)((1.0 / 60.0) * 1000.0);
     
     /**
      * How quickly the graphics are updated. 1 over the number of frames per second.
      */
-    const double FRAMERATE = (1 / 60.0) * 1000;
+    const unsigned int FRAMERATE = (int)((1.0 / 60.0) * 1000.0);
 
     /**
      * Boolean that represents whether the game should exit or not.
